requirements.txt
numpy==1.26.4
pandas==2.2.1
mysql-connector-python==8.0.20
psycopg2==2.9.9